# app/api/titles/route.ts
import { auth } from "@/auth";
import { fetchGenres, fetchTitles } from "@/lib/data";
import { NextRequest, NextResponse } from "next/server";

/**
 * GET /api/titles
 */
export const GET = auth(async (req: NextRequest) => {
  try {
    //@ts-ignore
    if (!req.auth) {
      return NextResponse.json(
        { error: "Unauthorized - Not logged in" },
        { status: 401 }
      );
    }

    const {
      user: { email }, //@ts-ignore
    } = req.auth;

    const params = req.nextUrl.searchParams;
    const page = params.get("page") ? Number(params.get("page")) : 1;
    const minYear = params.get("minYear") ? Number(params.get("minYear")) : 0;
    const maxYear = params.get("maxYear")
      ? Number(params.get("maxYear"))
      : new Date().getFullYear();
    const query = params.get("query") ?? "";

    // Debugging
    // Handle genres parameter
    let genres;
    const genresParam = params.get("genres");
    if (genresParam) {
      try {
        genres = JSON.parse(genresParam);
      } catch {
        genres = genresParam.split(",");
      }
    } else {
      genres = await fetchGenres();
    }

    const titles = await fetchTitles(
      page,
      minYear,
      maxYear,
      query,
      genres,
      email
    );

    return NextResponse.json({
      titles: titles,
    });
  } catch (error) {
    console.error("Error fetching titles:", error);
    return NextResponse.json(
      { error: "Failed to fetch titles" },
      { status: 500 }
    );
  }
});